+++ config/events.php
@@ -180,7 +180,6 @@ return function (App\EventDispatcher $dispatcher) {
             App\Radio\AutoDJ\Annotations::class,
             App\Radio\Backend\Liquidsoap\ConfigWriter::class,
             App\Sync\Task\NowPlayingTask::class,
-            App\Webhook\Dispatcher::class,
             App\Controller\Api\NowplayingController::class,
         ]
     );